migrated pop ups to slide
